fix: zoom out via slider
- fix: when zooming out via the slider for >1 step at a time, the
  un-zoom function encountered some undefined's
- add shift-mouse-wheel zooming
